requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
pydub==0.25.1
torch==2.0.1
torchaudio==2.0.2
python-dotenv==1.0.0
zipfile36==0.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
Flask-Migrate==4.0.4
SQLAlchemy==2.0.15
email_validator==2.0.0
